Integrated Spacecraft class
I changed the Spacecraft class to have a starting speed value, which is 0, and then, when launched, go to 1.
Also, I created the moveForward function which manages to move the ship in any sense (including side movement).
The Spacecraft now also knows about his starting point and his ending point as an ObjectLocation, interface related to locations.ts.