KeyListener class handling keypresses
I created this class which handles keypresses using the readline package (which was already the standard in this project).
Keypresses are handled and bound to events which can be added using the method bindEvent, or be removed using unbindEvent.
Events do have handlers and those handlers are executed by the KeyListener whenever the Event key is pressed.