CLI class to make all classes interact
This CLI class acts as a glue between the other classes (KeyListener and Spacecraft) to make the Spacecraft interact on keypress.
The CLI just creates a KeyListener and binds movement events, but also logs the first messages of spacecraft launch.
I created the basic events (W, A, S, D) and bound them to the spacecraft movement, but I also created a "close" event (with Ctrl+C or Ctrl+X) which can close the program.